feat(keepEmptyGroups): Add app configuration parameter to keep empty groups

+++ lib/GroupManager.php
@@ -126,9 +126,16 @@ class GroupManager {
 			return;
 		}
 
+		// keep empty groups if general-keep_groups is set to 1
+		$keepEmptyGroups = $this->config->getAppValue(
+			'user_saml',
+			'general-keep_groups',
+			'0',
+		);
+
 		if ($this->hasSamlBackend($group)) {
 			$this->ownGroupBackend->removeFromGroup($user->getUID(), $group->getGID());
-			if ($this->ownGroupBackend->countUsersInGroup($gid) === 0) {
+			if ($this->ownGroupBackend->countUsersInGroup($gid) === 0 && !$keepEmptyGroups == '1') {
 				$this->dispatcher->dispatchTyped(new BeforeGroupDeletedEvent($group));
 				$this->ownGroupBackend->deleteGroup($group->getGID());
 				$this->dispatcher->dispatchTyped(new GroupDeletedEvent($group));
